Change ELO storage from long to int
"int" should be enough for ELO score.  Plus it was the only place where a
"long" was used instead of an "int".

+++ src/delta.h
@@ -11,7 +11,7 @@ struct delta {
 		char *name, *clan;
 		long delta;
 		long score;
-		long elo;
+		int elo;
 	} players[MAX_PLAYERS];
 };
